Add 2.1 tag to image-inspector via 3.9 branch
(v3.* tags for image-inspector matching openshift versions are an accident
of using openshift release pipeline; it's not consumed as part of
Openshift, and has no separate versions per openshift release.
image-inspector has independent 2.* versioning.
It's consumed by Cloudforms which currently pulls 2.1 tag on all versions
of Openshift.)

https://bugzilla.redhat.com/show_bug.cgi?id=1642039

The use of additional_tags for image-inspector was started in
https://github.com/openshift/enterprise-images/commit/dc589a01acbd92da73b5b72ae9f4c3b1f9d20f76/#diff-e82d562d0ff7b956908e032665a888a2R14