images/ose-ovn-kubernetes.yml: re-enable, adjust repos
This is in the payload and at the moment is still in our spreadsheet of
things to ship, bugs claiming otherwise notwithstanding. Going to
reinstate it until I hear otherwise.

Also removed the fast-datapath-candidate repo. We should not be using
that for anything anymore. The image will need a rebuild.