Adding repo for OSP16 and switching ironic images
We'd like to start building all ironic images that will be shipped
in 4.3 using the content from Train release aka OSP16.
To be able to do that, this patch adds the OSP16 from the compose
passed_phase2 packages and switch all the related images to it
instead of the beta repo.
Leaving the beta repo untouched for the moment until we're sure
we can get rid of it.