non_release: clean up, add some OLM operators
as best I can tell, this only prevents the dist-git names from being
included in the images synced to quay and the release image.

i removed a lot of crufty entries that were no longer even configured,
and added a few that shouldn't be synced. nothing terrible happens if
those are synced, but they show up in a release-controller warning.