[ART-2065] ose-jenkins-agent-nodejs-12: put nodejs version in name
This was renamed from jenkins-agent-nodejs-12-rhel7 to
ose-jenkins-agent-nodejs, dropping the version.
Meanwhile ose-jenkins-agent-nodejs and component/comet repo is
deprecated (referred to unsupportable nodejs 8).
Moved to name ose-jenkins-agent-nodejs-12 as intended per
https://projects.engineering.redhat.com/browse/CLOUDBLD-1986
but leaving the internal image name un-versioned so in the CVO this
takes over from nodejs 8 as "the" jenkins nodejs agent.