Add jenkins nodejs{10,12} images to payload
I believe we intend to ship these two images as part of the payload, and
not as part of the extras. There is no operator governing these jenkins
agents.

As a background, these agent images were introduced after the GA. And
apparently did not make it in the payload, as we have previously based
our payload images on the previous payload images, which later got
codified with PR #559.

I am not sure if this is the right action to take in this phase of the
release, so consider this a conversation starter.