Add RHEL 7 pulp repo to get signed content
RHEL 7.9 initscripts, which is a dependency of s390utils, added a
dependency on bc.  While s390utils and its dependencies is removed from
the base image kickstart, bc wasn't added to the list of removals.  The
result is a signed bc ending up in the base image only on s390x.  When
bc is then installed in golang-builder from the brew buildroot which is
unsigned, a mismatch results and the build fails.

Adding a RHEL pulp repo, which is signed, and making it first
alphabetically should give it priority in yum mitigating the mismatch.