add fast-datapath repo for aarch64
We're trying to use the FDP repo for aarch64 to enable RHCOS builds on
that arch.  The usual repo sync via the release-controller doesn't
seem to be doing the trick, so it was suggested using this method.

See: https://github.com/openshift/release/pull/17851
See: https://github.com/openshift/os/pull/541#issuecomment-823601083