Pin the kernel for RHCOS 4.7 using RHEL 8.3
Until some point after OCP 4.8 GA, we have pinned RHCOS 4.7 to use
RHEL 8.3 content.

In order to not break `driver-toolkit`, let's pin the kernel versions
that are expected to be used during this time.

We don't anticipate any changes to which kernel is used unless there
is a CVE or urgent problem that affects OCP which needs to be
backported.