[ART-3892] update to use go 1.18 builder
Previously this was available but not used. With the obvious problems
solved, we are moving everything over to this go 1.18 builder.

The go 1.17 builder remains, just in case something needs to revert to
it for a temporary workaround, but that's not expected to happen. It can
be removed at code freeze.