Move rebase.sh invocation from Dockerfile to microshift-rebase
This commit moves the invocation of ./scripts/auto-rebase/rebase.sh from the Dockerfile to the microshift-rebase script. The change ensures that the rebase.sh script is run within the Podman container.

This change corrects an oversight in a previous commit where the invocation was removed from the Dockerfile but not added back in the microshift-rebase script.