Pin dtk until newer rhcos
rhcos builds are failing, and the consistency requirement that dtk and
rhcos have the same kernel build prevents us from creating nightlies.

This commit pins dtk to the build of the most recent nightly.

Note that nightly creation will fail again when rhcos builds succeed.
Which is what we want, because then we will not forget to revert this.