4.16: Revert default golang to 1.20
We currently have `canonical_builders_from_upstream: False` in
group.yml. This means that the golang builders defined in ocp-build-data
for images are taken, without consideration of what builder images are
defined in upstream dockerfiles.

Recently, while adding the 1.21 golang, we made it the default. This
means that the production builds made the change immediately.

To set `canonical_builders_from_upstream` to the value we'd expect at
this phase of 4.16, is a bit risky. Image builds will potentially break
for those images that made the switch to rhel9, but have not had that
switch reflected through reconciliation PRs.

The proposal of this PR is:
- to restore the default golang builder to 1.20
- to work to get existing reconciliation prs merged, in order to switch
  canonical_builders_from_upstream to true
- if anyone needs 1.21, open a PR to this repo for your image to consume
  the 1.21 builder