Remove olm image from any dependents
`dependents` signifies that the dependent image needs to rebuild when
this image changes. (This mechanism exists on top of rebuilding when a
base image changes). This is useful for operators, as they need to get
rebuilt with the updated image references. As olm is a payload image,
and thwe images it depends on as well, there is no need to rebuild the
olm image when an image this paylooad operator depends on, changes.