Fix rhel layers oc-mirror-plugin
After https://github.com/openshift/oc-mirror/pull/804 got in, production
build configuration followed up with https://github.com/openshift-eng/ocp-build-data/pull/4475.

This was incomplete, as the change to the final layer was not reflected
in this configuration update.

This PR sets rhel layers as upstream expects, and ensures ENTRYPOINT
does not run into a linking error.