Fix ci-build-root build failures
Brew builds of ci-build-root rhel8 images are failing the content set
check of brew, as the package less is different between the arches. This
is because less gets removed from some of the arches when git gets
installed. Reinstalling git pulls in the current less on arches where it
was removed, whereas it remains the old one where there was no removal.

This change instructs dnf to not be smart and remove perceived
dependencies.