Ignore stop-ship-tag for pre-GA releases
If the stop-ship tag is applied to a kernel, that is a signal that OCP
should not be shipping that release. Strictly speaking, this does not
apply to releases that are not shipping yet. CI of pre-GA releases
should be informing such a decision.

This is a minimal change to not let auatomation run havoc on pre-GA
releases, and to make it a human decision to retract the kernel or not.