ci build root images: No nobest
With `--nobest`, packages that are explicitly being installed, but
where an outdated version is already installed, will not see an
upgrade. This causes problems for packages that are available on, say,
x86_64, but not on one of the other arches. The OSBS plugin that checks
for consistent RPMs across the arches will fail the build.

This PR says: _Do better, dnf!_