csi-node-driver-registrar: Consolidate delivery repos
csi-node-driver-registrar is used in the payload, and in ART managed operators, and in external operators.
For the rhel8 versions, both ART and external operators refer to the image as csi-node-driver-registrar,
without the `-rhel-8` suffix.
This change consolidates delivery repos on what is actually used.
The operators referencing the rhel-less name (ocs-client-operator also in 4.18 iib):
- aws-efs-csi-driver-operator
- gcp-filestore-csi-driver-operator
- lvms-operator
- ocs-client-operator
- ocs-operator
- secrets-store-csi-driver-operator