csi-external-provisioner: Consolidate delivery repos
csi-external-provisioner is used in the payload, and in ART managed operators, and in external operators.
For the rhel8 versions, both ART and external operators refer to the image as csi-external-provisioner,
without the `-rhel-8` suffix.
This change consolidates delivery repos on what is actually used.
The operators referencing the rhel-less name (ocs-client-operator also in 4.18 iib):
- aws-efs-csi-driver-operator
- gcp-filestore-csi-driver-operator
- lvms-operator
- ocs-client-operator
- ocs-operator