ose-csi-external-resizer: Consolidate delivery repos
ose-csi-external-resizer is used in the payload, and in ART managed operators, and in external operators.
For the rhel8 versions, both ART and external operators refer to the image as ose-csi-external-resizer,
without the `-rhel-8` suffix.
This change consolidates delivery repos on what is actually used.
The operators referencing the rhel-less name (ocs-client-operator also in 4.18 iib):
- gcp-filestore-csi-driver-operator
- lvms-operator
- ocs-client-operator
- ocs-operator